Fix Rpl_semi_sync_master_wait_no_slave description
The Rpl_semi_sync_master_wait_no_slave description could be interpreted
as _all_ transactions would wait for a semi-sync ACK when
Rpl_semi_sync_master_wait_no_slave is ON with no replicas connected,
indefinitely. This is incorrect, as it is only the next transaction
which commits (and any others which commit afterwards, up through
the semi-sync timeout), and then the primary will revert to
asynchronous replication.

Corrected the description to be more clear to the underlying
behavior.