MXS-5904: Update WCAR install docs
The visualization needs python 3.9 and on all systems except RHEL 8 it's already the default. Added a note about the potential devel package dependency and clarified the RHEL 8 installation instructions.

Copied over the newest version of the WCAR visualization documentation that also includes the commands needed to set up the visualization on a non-localhost address.