Remove legacy/confusing read-only-replicas documentation
The bullet points for fixes on read-only-replicas from 10.3.19 is
confusing, as it isn't clear as to what behavior was incorrect and
what the correct behavior should be. 10.3.19 is also EOL, so the
fixes are not as relevant now anyway. The text above these issues
describes more generally what is correct, so I think it would be less
confusing to just remove the snippet about the EoL fixes.

Note MDEV-21442 is a report that directly stems from this confusion,
and comments on that JIRA show another user has been confused by this
as well.